+++ src/test/java/dk/kb/present/transform/XSLTPreservicaSchemaOrgTransformerTest.java
@@ -589,7 +589,7 @@ public class XSLTPreservicaSchemaOrgTransformerTest extends XSLTTransformerTestB
     @Test
     public void noGenreTest() throws IOException {
         String transformedJson = TestUtil.getTransformedWithAccessFieldsAdded(PRESERVICA2SCHEMAORG, TestFiles.PVICA_RECORD_03f18f50);
-        assertTrue(transformedJson.contains("\"genre\":\"Radio-rodekasse\""));
+        assertTrue(transformedJson.contains("\"genre\":\"Blandet Radio\""));
     }
 
     @Test