+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Marked unittest with @Tag("integration") so it can build without aegis. 
+- Rename genre 'TV rodekasse' to 'Blandet TV' and 'Radio rodekasse' to 'Blandet radio'
 
 ## [4.0.0](https://github.com/kb-dk/ds-present/releases/tag/ds-present-4.0.0) 2026-01-29