+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to ds-present will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.1](https://github.com/kb-dk/ds-present/releases/tag/ds-present-4.0.1) 2026-03-04
+
+### Fixed
+
+- Handle datetime with milliseconds using OffsetDateTime
+
 ## [4.0.0](https://github.com/kb-dk/ds-present/releases/tag/ds-present-4.0.0) 2026-01-29
 
 ### Added